use BigInt instead of PositiveInt for upload offest field
The regular int can overflow with large but reasonable file sizes.
Note that BigInt provides no assurance that a value is positve,
which is unfortunate, but PositiveBigInt is not currently a supported
field in django.

+++ drf_chunked_upload/models.py
@@ -34,7 +34,7 @@ class ChunkedUpload(models.Model):
     user = models.ForeignKey(AUTH_USER_MODEL,
                              related_name="%(class)s",
                              editable=False)
-    offset = models.PositiveIntegerField(default=0)
+    offset = models.BigIntegerField(default=0)
     created_at = models.DateTimeField(auto_now_add=True,
                                       editable=False)
     status = models.PositiveSmallIntegerField(choices=STATUS_CHOICES,
